full redesign
I need to delete files and recreate the react app with updated node and npm packages in order to develop a new site. would making a new repo be easier. Yes. But then i'd have like three "portfolio" repos, and this one is pinned so I'm using this one.
So there.