rebuilding the site
The NPM packages were all out of date and we were running on an old version of node. To fix this, I'm going to recreate the react app and paste in the src that's left. Then I can see what I'm working with for a portfolio (Which... won't be much and I'm just going to redo it all anyway but at least there'll be something there)